Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "eval-type-backport>=0.2.2",
    "fastapi[standard]>=0.116.1",
    "numpy>=2.3.1",
    "openai-agents>=0.2.3",
    "pillow>=11.3.0",
    "presidio-analyzer>=2.2.359",
    "presidio-anonymizer>=2.2.359",
    "pypdf2>=3.0.1",
    "python-doctr>=1.0.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "uvicorn>=0.35.0",
]
